feat!: initial invites implementation
Implements only invite creation and deletion. Invite invokation, "joining a guild," will be added later on alongside fetching invites.

Massive amountrs of refactoring will need to be done with the Protobuf in charge of sending events to the Gateway since we now use Rust. I cannot for the love of god find anything to simply convert a Serde struct to a HashMap in Rust, so best case is for now until it's a limitation to just use JSON on both sides of the conversation until we have something better.