feat!: gateway v2
A refreshed and rethought Elixir Gateway. The Rust Gateway kept being too much of a challenge to undertake and would have been a huge maintenance burden due to the many more failover systems needed and the complexity compared to Elixir. This is unfinished, and at the moment doesn't work, but I just wanted to commit something to show actual progress is, well, happening.